Fix links to NOOP-translation functions for QML
qsTr*NoOp functions are not documented, are not registered in
qqmlbuiltinfunctions.cpp, and are not tested or otherwise exercised
anywhere in our code base. I conclude that they don't exist, and that
the correct functions to use are the macro-like QT_*_NOOP functions.

Task-number: QTBUG-124657
Change-Id: I26632eae32b16a15617b095c5ea116b34bdbcb2e
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 98eb19963148e6dc4397fae7d1f041290336528d)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>