CMake: Alternative fix of deployment for lightningviewer example
qmlimportscanner is used to scan the qml files, to decide which
plugins and libraries to deploy. Because none of the qml files
explicitly reference the LightningViewer or LightningViewerStyle
imports, it never deploys the qml modules, nor their dependencies.

This change adds a Deployment.qml file to the sources, which is not
actually used at runtime, but because qmlimportscanner sees it, due to
scanning the whole project directory, it will deploy the necessary
files.

Pick-to: 6.7 6.8
Fixes: QTBUG-127835
Task-number: QTBUG-128225
Change-Id: Ic5185d723d3a1f65427af1ee7799fa49ea414cfd
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: MohammadHossein Qanbari <mohammad.qanbari@qt.io>