Doc: Restructure deprecated module section in What's New in Qt 6.8
The section contains a list of two elements that are arguably a list
of one, where the second item belongs with the text that preceeds
it. Turn the list item into a subsection and touch the text slightly
such that it reads better.

Change-Id: Iaf6cd92991126ad13d738c94f9bd78a77a9e5151
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
(cherry picked from commit 1a720cc041d35cf658f52928bbdf9e630977a8e3)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>