Doc: Add link to "What's New" page from prominent location in index
Make use of the `\QtVer` macro, which expands to the major.minor
version of Qt that's configured when building the documentation. This
way, the link should remain up-to-date and not need manual
maintenance. Note that this will cause QDoc to emit a warning about
failing to link to a version if that version of Qt doesn't contain a
page with the title "What's New in Qt VERSION", where `VERSION` is the
string `\QtVer` expands to. This will always be the case when bumping
version numbers, until the new page is added. However, this warning is
ignored in COIN and should thus not cause integration issues (but
rather serve as a reminder that such a page is missing).

Fixes: QTBUG-128399
Pick-to: 6.8
Change-Id: I274d56d3d0b7edd50c369608a2710749f35219b2
Reviewed-by: Andreas Eliasson <andreas.eliasson@qt.io>