Restore support for iOS 16 for Qt 6.8
Our general strategy is to support the latest two iOS releases,
as most people upgrade their iOS devices quite aggressively.

But given that iOS 18 is not out yet, and that people still
spend some time upgrading once it's out, we don't want to be
in a situation where we effectively only support iOS 17.

Task-number: QTBUG-128678
Pick-to: 6.8
Change-Id: Ic2c9da1e3775ff1e134b50c63410a61a1bfeaf36
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>