Disable CarRendering example on Yocto
Car Rendering example relies on CMake's FetchContent to fetch
the Qt Design Studio Components it's using. But for a normal
BitBake/Yocto build, FetchContent is disabled, and will result
in weird configure errors.

The patch detects a Yocto build by checking for the BBPATH
environment variable. If set, the QT_EXAMPLES_WITH_FETCHCONTENT
cache variable is set to false, and the Car Rendering example
will be excluded from the build.

Pick-to: 6.8 6.8.0
Fixes: QTBUG-128729
Change-Id: I17783aff00f996f6c35fc1b73d69ef355c14c2ee
Reviewed-by: Jani Heikkinen <jani.heikkinen@qt.io>