Doc: All Qt modules: List all addons in a single list
Stop separating GPLv3-licensed modules in their own list, to make it
easier to locate individual addons.

Instead, list them in the licensing documentation. For now, use a
hardcoded list. Eventually this should be replaced with automatic
grouping, with \ingroup commands in the source for each Qt module's
docs.

Make the list of addons directly visible on the Qt landing page,
instead of using an expandable element that hides them by default.

Fixes: QTBUG-129132
Pick-to: 6.8 6.8.0
Change-Id: I980934640ca1c5a24d192061701c1fee91964cf7
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>