Docs: Improve the root tree structure
This commit reorganizes the root tree structure. The new "Product
Information" subtree includes subtrees about licensing, supported
platforms, and releases. The subtree about building UIs with Quick is
moved to overviews and will sunset in the nearest future. Its content
will be moved to Qt Quick.

The References subtree was moved to the topmost position because of two
reasons: it keeps the existing doc usage workflow possible and the
reference section does not feel suitable at the bottom of the curent
root tree. It happens because after the main content of the root tree,
the module subtrees are attached. It's strange to see common concepts
and modules separated by API references.

Task-number: QTBUG-133740
Pick-to: 6.9
Change-Id: I6e9aa0e6cb2cc4b07fed8d3bf5e146aee9b3df67
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>