Dice: Improve atRest property of PhysicalDie
atRest property should be true while the motionTimeout timer
is running.

This change remove a function binding and directly aliases the
atRest property to running property of the timer.

TODO: atRest is unused based on a previous change. If we decide
to keep the logic in the previous patch, then we can remove
this timer and also the property.

Pick-to: 6.8 6.9 6.9.0
Change-Id: I9d0c58f192251621cb3e691cf431366eb6b1dd68
Reviewed-by: Nicholas Bennett <nicholas.bennett@qt.io>