Doc: Prevent main content from overflowing Contents side section
Having a \table with 3 \inlineimages on each row is prone to cause
overflow issues on narrower screen widths. Listing the three groups
of \inlineimages together with a \section1 and a \br is more dynamic
and will wrap the images as the screen width changes.

Fixes: QTBUG-134789
Pick-to: 6.8 6.5
Change-Id: Id05df222fea226e602a74b57bef0bce0a55b86a3
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
(cherry picked from commit 88402b30620764af3de401143231d34e0613df1f)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>