Doc: Remove empty "Manual" hierarchy item from "Build with CMake"
In the "Contents" tree view in Qt Assistant and in the side bar on
doc.qt.io was an additional "Manual" item just below "Build with CMake"
that contained the links to all pages of the manual. That looked out of
place compared with the siblings of "Build with CMake".

We now change the type of the subproject "manual" to "manual". The
selector seems to be ignored, and we remove it.

Caveat: We have to remove links to sections from the table of contents
in "Build to CMake", because
- links to sections are dysfunctional in Assistant's "Contents" tree
- Assistant renders "\tt Qt6::Core" as "typewriter Qt6::Core"

If the "fake:page" selector wasn't ignored for the "manual" type then
this stunt wouldn't be necessary.

Change-Id: Ib58b6c432ad2b543d2c8a840906951884f9416cb
Reviewed-by: Alexei Cazacov <alexei.cazacov@qt.io>
Reviewed-by: Kai Köhne <kai.koehne@qt.io>
(cherry picked from commit b2f94213ba8abbd904a285a47b5d01c739384fd7)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>