Doc: Move "Importing Qt" section out of "Getting started with CMake"
The information in the "Importing Qt" section is too detailed
for the introductory nature of the "Getting started with CMake" page.
Move the content to a separate page and fix some issues with macro
usage.

The page is in a separate file now to make it easy to move it around in
the documentation hierarchy.

The phrasing "Importing Qt" is a bit unusual. Change it to the longish
but clearer "Make Qt available".

Change-Id: Ia23a3870eb82407d3652a95d91e99599c2437678
Reviewed-by: Kai Köhne <kai.koehne@qt.io>
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>
(cherry picked from commit 7aaed26a0bee3fabef11041636c3b93a61d712bf)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>