Adjust delegate width to fit the URL in colorpalette
The server URLs got an additional /API path and they didn't fit into the
delegates anymore. This patch increases the delegate width so it looks
nice again. This is not the best solution. A dynamic solution would be
better but also a bit too complicated for an example.

Amends 6a6212a65289fdccefa565d39eb74e31761107bf.

Pick-to: 6.9 6.8
Change-Id: Iead827d4e37b5e0494f7f35c95934b8bf848cb40
Reviewed-by: Juha Vuolle <juha.vuolle@qt.io>