Add api-key hearer to colorpaletteclient example
Colorpaletteclient is able to interface two server backends: public
reqres.in service and QHttpServer example service. The public reqres.in
service has recently started requiring an api-key HTTP header (see
https://reqres.in/signup).

This commit adds an api-key header if the app uses the public
reqres.in service. The header is added to the QNetworkRequestFactory
and is consequently then used in all subsequent requests.

The code is furthermore  adjusted so that it won't accidentally
remove the api-key when toggling another proprietary header ("token").

Pick-to: 6.8
Fixes: QTBUG-136790
Change-Id: Id19d223722fd8f735d76855a293bc77f680ed344
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>
(cherry picked from commit 01a52994531554005769c34be8fdc28ee70a7386)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>