Docs: Remove FAQ section, integrate/update the note about fragments
FAQs are not the best practice for technical writing.

The points were moved to the relevant Qt for Android pages.

The fragments point was moved to the top-level Qt for Android Page, and
updated to reflect that you can use Qt Quick for Android Views inside
a fragment in an Android Application.

Task-number: QTBUG-137077
Pick-to: 6.9 6.8
Change-Id: I3a35f1c80cb5036822e025f1164f4c1414c85454
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>
(cherry picked from commit 0d6609677d909e06c94631a371f9eff3074297d7)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>