Doc: Do not treat the 'Getting Started Programming with Qt Quick' tutorial as an example
This tutorial is so far marked and treated as an example, and shows up e.g.
in the Qt Creator Welcome screen. However, the documentation content and title makes it
clear that it is actually a tutorial. Treat it as such.

Task-number: QTBUG-137988
Change-Id: Icb30ef31956a5e26b3f9c928245eb533c093272d
Reviewed-by: Alexei Cazacov <alexei.cazacov@qt.io>
Reviewed-by: Nicholas Bennett <nicholas.bennett@qt.io>
(cherry picked from commit 1a1b8765e7d5b408743be0a8bc9a6ef2111b02b7)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>