Doc: Use sort keys for What's New documents to ensure correct sorting
The existing behavior for sorting in QDoc does not take into account the
way version numbers are ordered, so 6.10 is treated as a lower value
than 6.2, for example. The workaround for this is to provide sort keys
that specify 6.10 as 610 and 6.2 as 602. Similarly for the other version
numbers.

Fixes: QTBUG-137956
Change-Id: Icddd4fef110008fd208d3a564cad7ad511ac4fe2
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
(cherry picked from commit 0d757db0c8c25f7df97c44a4f8fca034c83b5681)