Coffe Machine: Configure qmllint
We're at the point where we can enforce 0 warnings, if we decide to
explicitly ignore the custom parser one (which we need to ignore at
least for the ui.qml files).

Running the all_qmllint target will now fail the build if a new warning
is introduced (but the target is currently not run by default).

Pick-to: 6.10
Change-Id: I60a232dc838574701c0f5a9db9245ff0f8f24525
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>
Reviewed-by: Luca Di Sera <luca.disera@qt.io>