coffee example: Clean up dataflow (1/n)
Make the dependency on AppicationFlow explicit in Settings by using a
required property. Use that to move the binding to ApplicationFlow's
properties into settings itself, instead of setting them up in the
"settings" component – this reduces the dependency of
ApplicationFlowForm on its derived type. Initialize the required
property when pushing the component on the StackView.

As a drive-by, use the more modern pushItem function instead of push.

Pick-to: 6.10
Task-number: QTBUG-138164
Change-Id: I47f2931d57ccd66c47775923a15e23d371afa6d1
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>