coffee example: Clean up dataflow (2/n)
Make the dependency on AppicationFlow explicit in Progress by using a
required property. Use that to move the binding to ApplicationFlow's
properties into Progress itself, instead of setting them up in the
"progress" component – this reduces the dependency of
ApplicationFlowForm on its derived type. Initialize the required
property when pushing the component on the StackView.

As a drive-by, use the more modern pushItem function instead of push.

Pick-to: 6.10
Task-number: QTBUG-138164
Change-Id: I9e483c97b7585cef2fed043aa3a5ab35a4b54b10
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>