coffee example: Clean up dataflow (4/n)
Deal with Home. Home differs from the other pages, because it is used as
the initial item, and thus an actual instance is already created in the
ui.qml file, not only a Component.
As we still want to use a required property referencing the
ApplicationFlow instance, the instantiation is moved into
ApplicationFlow (instead of ApplicationFlowForm).

Task-number: QTBUG-138164
Change-Id: I903352e17560c056fdb12e370ee5413adbe2fd1d
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>
(cherry picked from commit daf3f87430363964b85438d58bc754060d40972d)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>