robotarm: Port QML visible property to QObjectBindableProperty
This allows us to give it a signal, which silences a qmllint warning.
While the engine does not actually need that currently, it is planned
that we want to slowly get away from supporting non-notifying bindable
properties.

This allows us to set max-warnigs to 0 for qmllint.

Fixes: QTBUG-138176
Change-Id: I3221095f31639d90ebcc9d54fd4376cfb9140961
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>
(cherry picked from commit 0da5048dd7ed930ef2e83ba12dc8d67dcc39fb09)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>