robotarm: clean up main
Use loadFromModule, instead of load, as it's the simpler and newer API.
That requires renaming main.qml to Main.qml, and some changes to the
CMakeLists.txt file. That gets rid of a runtime warning about using
relative URLs.

Moreover, avoid adding an import path we don't need, and remove the
error handling for the case where loading fails – we know it can't fail.

Pick-to: 6.10
Fixes: QTBUG-138177
Change-Id: I01a3eac70147bf5aef6456be0b634af933daae20
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>