ToDo example: Avoid calling functions in ui.qml files
QDS does in general not allow function calls outside of Connections.
We could either rename the files to plain .qml to acknowledge that, or
use alternative constructs.
This patch opts for the latter aproach: We use Connections in most
places, and expose the subcomponent in another (so that the logic can be
handled in the already existing non-ui.qml file).

Pick-to: 6.10
Task-number: QTBUG-138189
Change-Id: I27a5cc6fc25de73faddeeaa2e103f4df03c5c18f
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>