Doc: Remove 'Solutions for UI Design' page
The Doc team is currently improving the onboarding documentation for users.
As per this improvement, the consensus is to focus on Qt Quick, especially for
new users, and to remove information and comparisons to widgets in the
Getting Started docs.

This patch removes this page as it introduces both widgets and Quick to new users.
This page is also removed from the TOC.

Fixes: QTBUG-138445
Pick-to: 6.10
Change-Id: Iaf20466257859ca76140127f6f97023eef5ff062
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>