ToDo example: Fix remaining unqalified access
SettingsViewForm interacts with the "global" stack view to push
different (sub-)pages on it. Passing that explicitly down to it would
look somewhat weird, because it would pass through components which
don't directly need it.
So instead, still keep the very tight coupling, but ensure that we at
least have type information: We can get access to the view's Window via
Window.window; and can cast that Window to App. From App, we can then
expose the StackView as an alias property.

A more ideal solution might be to decouple the logic of determining the
next page from actually navigating there. One could for instance add a
signal "transitionToView(page: Component)", and then handle the
navigation on a higher level. However, that would necessiate larger
changes, and is therefore not implemented here.

Pick-to: 6.10
Task-number: QTBUG-138189
Change-Id: Ie352c5afc3a9bd3b2c2a6fc9acdca2d23bbcadbf
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>