Doc: Update 'Revise Your First Applications' page
The Doc team is currently improving the onboarding documentation for
users. As per this improvement, the consensus is to focus on Qt Quick,
especially for new users, and to remove information and comparisons to
widgets in the Getting Started docs.

This patch does the following:
- Removes the Widgets Tutorial and Examples section from
the table.
- Fixes the link to the Mobile examples
- Removes the sentence referencing the MS Visual Studio, as it is does
not fit into the context.

Fixes: QTBUG-138449
Pick-to: 6.10
Change-Id: I2748e051e6ec16662ba94e194141a10a0a0a3d2d
Reviewed-by: Kai Köhne <kai.koehne@qt.io>