Doc: Remove 'Tools for Qt WIdgets UIs' page
The Doc team is currently improving the onboarding documentation for users.

As per this improvement, the consensus is to focus on Qt Quick,
especially for new users, and to remove information and comparisons to widgets
in the Getting Started docs.

This patch does the following:
- Removes this page as it introduces Qt Widgets Designer to new users.
- Removes links to this page
- Removes qtdesigner.png image

Fixes: QTBUG-138448
Change-Id: Iec7f4c819aff6bd77a4cf4f4492b6276fc52d451
Reviewed-by: Kai Köhne <kai.koehne@qt.io>
(cherry picked from commit ec2535e06cbef39cdf897860b06cc82a76eb89de)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>