MediaPlayerApp: Fix qmllint [import] warning
`MediaPlayerApp` depends on some additional custom QML modules that are
positioned such that they are siblings to each other.

By default they would not appear under the normal QML search path, such
that qmllint was failing to find them producing a series of [import]
warnings.

Add a direct dependency, based on CMake targets, to the custom
`Config` and `MediaControls` QWL modules to ensure that qmllint finds
them, thus fixing the warnings.

The `add_subdirectory` calls in the main `CmakeLists.txt` file of the
project were re-ordered to ensure that the dependencies targets are
known at the time we add them as a dependency.

Task-number: QTBUG-138175
Pick-to: 6.10
Change-Id: I5ce43606ae0458a572b97b6afb218aafa8ef87b8
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>