MediaPlayerApp: Fix [unqualified] warnings in MediaControls
The `MediaControls` QML module that is part of the `MediaPlayerApp`
example, imports a custom QML module, `Config`, that is defined in a
sibling directory.

The `Config` module exports a singleton and, when the singleton is
accessed, qmllint produces a series of [unqualified] access warnings.

Add a direct dependency to the custom `Config` QWL module to
ensure that qmllint recognizes the singleton, thus fixing the warnings.

To allow for the use of the modern, `TARGET` based, dependency
declaration method, introduced in 6.8, the `find_package` and
`qt_standard_project_setup` calls for the project were modified to
search and setup for the 6.8 version.

The `add_subdirectory` calls in the main `CmakeLists.txt` file of the
project were re-ordered to ensure that the `Config` target is known at
the time we add it as a dependency.

Task-number: QTBUG-138175
Change-Id: I33cbbdd179d701a36bb5c4f38deca1102aa47a90
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>
(cherry picked from commit a2a8d24d50d647dfe3e8d669fc2f23d5286021e7)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>