MediaPlayerApp: Fix [unqualified] warning in the MediaPlayer module
The `MediaPlayerApp` example has has a `MediaPlayer` QML module whose
`Main.qml` file defines a function, `openFile`, in the scope of the root
element, that is later used by a child element in a binding.

qmllint produces an [unqualified] access warning for this usage.

To avoid the warning being produced, the usage of the `openFile`
function in the child element was qualified with the id of the root
element.

Task-number: QTBUG-138175
Pick-to: 6.10
Change-Id: I11f8c02fb2c45c178f42770cb06badb3fb73c3e1
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>