MediaPlayerApp: Add a configuration file for qmllint
The example is currently in a state where qmllint produces no warning,
and should generally be kept in such a state.

Thus, add a configuration file that sets the warning limit to "0", such
that running the all_qmllint target for the example will fail if any
warning is produced, with the additional strictness hopefully helping
keeping the example warnings-free.

Pick-to: 6.10
Fixes: QTBUG-138175
Change-Id: If9adb0f4929fdb99272883ceccca9479bf640204
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>