Thermostat: Update RoomsModel
Change the model from a ListModel to a list<Room> where Room is a new
type containing all the information related to a room. This allows
making what is passed to various components and delegates clearer and
explicit.

The data of the model itself was also updated:
-The statistics for humidity, energy, and temperature for each room are
 now stored as simple list<int> or list<real>.
-Data series were regenerated to be more varied; many were identical or
 didn't fluctuate enough to visibly change dynamically in the UI.
-More rooms were made active from the start and especially the first one
 which allows for immediate interaction in the thermostat settings view
 without having to first activate the room.
-Background values for humidity and energy consumption were also made
 more varied instead of identical everywhere.

The rest of the example will be adapted to this new model in the next
commits.

Task-number: QTBUG-138188
Change-Id: I3ce48ccb14aa4cfdee2d0508427b26598ce5446a
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit 147f78d7de550679e829e54671168c2c8e467a9b)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>