FX&MaterialShowroom: Access font trough the Application singleton
The `Qt` singleton has an `application` member that is not guaranteed to
be a `QQuickApplication`, and thus might not always have a `font`
member.

Hence, qqmlint can produce a [missing-property] warning when, for
example, an access like `Qt.application.font` is performed.

To avoid the issue, the `Application` singleton that QtQuick exposes can
be used for the same purpose, while ensuring that the singleton is
recognized as a `QQuickApplication` and has a `font` member always
available.

Hence, ensure that accesses to the `font` property always pass by the
`Application` singleton instead of the `application` member of the `Qt`
singleton.

Task-number: QTBUG-138170
Change-Id: I59a1c8307e7e8f5df6a65ba810eebcf2d8e37bba
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit eab7334e03a0115c8d18361ec28b07106e7e7dda)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>