FX&MaterialShowroom: Remove unused types from PocketDemo
The examples uses an internal QML module, `PocketDemo`, that provides
various QML types, most of which seems to be unused in the rest of the
project, and that further depend on dependencies that aren't currently
provided by the project itself.

Hence, remove the files defining those types.
Further adjust the CMakeLists.txt file and qmldir file for the
PocketDemo module to avoid referring to the now removed files.

Pick-to: 6.10
Task-number: QTBUG-138170
Change-Id: I4f76879f0c39652ccae4fac1e828bbf38c3a3c33
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>