FX&MaterialShowroom: Remove unused types from QtQuick.Studio.Components
The examples uses its own version of some of the designer studio
components that provides various QML types, most of which seems to be
unused in the rest of the project, and that further depend on
dependencies that aren't currently provided by the project itself or
produce various qmllint warnings.

Hence, remove the files defining those types.
Further adjust the CMakeLists.txt file and qmldir file for the
QtQuick.Studio.Components module to avoid referring to the now removed
files.

Pick-to: 6.10
Task-number: QTBUG-138170
Change-Id: I88d1e0e7bfa63a70d46c97e970e243a1883da84b
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>