FX&MaterialShowroom: Fix qmllint [import] warning in content
The various ui and qml files under content depends on some additional
custom QML modules that are part of the example, living under the
`imports` directory.

The module for those files does not link to those custom modules, such
that qmllint will produce a warning when they are imported.

Add a direct dependency, based on CMake targets, to the custom
`PocketDemo` and `CB_MaterialBundle` QWL modules to ensure that qmllint finds
them, thus fixing the warnings.

The `add_subdirectory` calls in the main project description were
reordered to ensure that the dependencies targets are available when we
declare a dependency to them.

Pick-to: 6.10
Task-number: QTBUG-138170
Change-Id: I743524dafec41923664fe45decd946237b6c21ba
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>