FX&MaterialShowroom: Set the ComponentBehavior to bound where necessary
Some of the (.ui).qml files access some elements by qualifying them with
the identifier of an outer components.

This kind of access requires the ComponentBehavior to be set to Bound,
with qmllint warning about it when this is not the case.

Thus, to fix the warnings, set the ComponentBehavior to Bound in those
files that make use of the relevant access pattern.

Pick-to: 6.10
Task-number: QTBUG-138170
Change-Id: Id39e07e25769c063b5a41d09632584b5c74a5fd8
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>