FX&MaterialShowroom: Fix [incompatible-type] warnings in content
Many of the ui.qml files were assigning single objects to properties
that expects a list of those objects.

For example assigning a single particle object to the `particles`
members of a `Wander3d`, with qmllint producing warnings about the
incompatibility of the binding.

To fix the warnings, change all of those patterns to wrap the single
object into an array.

Pick-to: 6.10
Task-number: QTBUG-138170
Change-Id: I371da8e70faa25bf009186c16fa722457eb372c5
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>