FX&MaterialShowroom: Fix some [unqualified] warnings in Screen01
qmllint was reporting an unqualified access warning to `pinch` in a
connection based onPinchUpdated handler of a PinchArea element.

The `pinch` in the relevant function is intended to be the `PinchEvent`
that is passed to the handler, thus name the first parameter of the
function to mark the argument explicitly which in turn stops qmllint
from warning about it.

Pick-to: 6.10
Task-number: QTBUG-138170
Change-Id: Iffb0ca2c4ff9a64cb2be880788055146083be728
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>