FX&MaterialShowroom: Fix some [unqualified] warnings in Screen01
qmllint was reporting an unqualified access warning to `pinch` in a
connection based onPinchUpdated handler of a PinchArea element.

The `pinch` in the relevant function is intended to be the `PinchEvent`
that is passed to the handler, thus name the first parameter of the
function to mark the argument explicitly which in turn stops qmllint
from warning about it.

Task-number: QTBUG-138170
Change-Id: Iffb0ca2c4ff9a64cb2be880788055146083be728
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>
(cherry picked from commit 57325f89495b6572de9ba002f13e77a92e43655c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>