Doc: Fix incorrect usage of \generatelist
The \generatelist command takes an optional argument
that affect how the members are listed, for example,
`classesbymodule` which creates an annotated list of
the classes in a named C++ module.

`groupsbymodule` is not recognized or documented. QDoc
falls back to generating a list of all members of a
group in this case. While the end result is correct,
this relies on an undocumented feature that is likely
to change and break.

Replace these instances with an explicit \annotatedlist
command that produces the same output.

Task-number: QTBUG-138901
Change-Id: I9de11926187daca8d30bf63d6ef8cc7bcded1e0d
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
(cherry picked from commit fcea1e1e87646a702c500558e3776782538ff57b)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>