FX&MaterialShowroom: Fix access to implicitly available elements
Across the example there are many accesses to implicitly available elements.
That is, an element is declared in some file with a certain id, and then
other files use that id directly, to access those elements.

While the engine can handle this, qmllint cannot, as it cannot know that
those accesses will happen in a context where the referred to elements
with the referred to id are available as that can only be resolve at
runtime.
Thus, it produces a series of [unqualified] warnings for those accesses,
as it cannot make sense of them.

To fix the warning, and generally improve the style/readability of the
example, the pattern for those accesses was modified in a way that can
be understood by the tooling.

In particular, whenever a custom type wants to access an implicitly
available element, it now provides a required property to an element of
that type that will be filled-in at usage-site, where the previously
implicit element is available and can be explicitly bound.

The various accesses to the objects or its subparts were updated to use
the newly added required properties and the various usage site of the
relevant types were updated to ensure that those required properties are
bound to a relevant element.

In some cases, properties that were defined inline without a backing
type were being accessed on an implicitly accessed element.
One such example is in the case of certain `NumberAnimation`s elements
being defined in "Screen01.ui.qml", with an additional "phase" property
being defined inline and then accessed through an implicit reference to
that specific `NumberAnimation` instance.

Since this is done without specifying a type, qmllint cannot be induced
to know about the presence of the additional property, such that it
might fail to understand the access to it.
In those cases, a required property was added that is bound directly to
the additional subpart of the element instead of the element itself,
which allows qmllint to understand the provenance of the property and
validate the access.

Task-number: QTBUG-138170
Change-Id: I05ce940e7de45b1749010af67fce2ac571cf9ba6
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>
(cherry picked from commit a8b7fdcbd666ff5b55e3b461c63a7ef631e93357)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>