FX&MaterialShowroom: Fix [unqualified] accesses in State_Idle
The example uses a custom QML type, `State_Idle`, to handle a series of
lists and their selections.

In particular, `State_Idle` is used to change the state of a string
property of each of those lists that represent the current selection.

This change is done by accessing a `list` element that is supposed to be
implicitly available and that is currently being defined inline when the
relevant list elements are created.

qmllint cannot know that at runtime `list` will be an available property
and thus produces a warning on access.

Generally, this type of accesses is fixed by explicitly passing the
dependency to the relevant element, that can then access it in a way
that qmllint can keep track of.

The subpart of those types that is modified by `State_Idle` is a string.
Passing a string downward would disallow `State_Idle` from changing the
original value, as the binding would only go one-way.

Furthermore, `State_Idle` is used for interacting with multiple types,
for example `ListModels` and `ListMaterial`.
All of those elements don't have a common shape that is known to have
the necessary property available, such that there is no single type that
can be used as the target of the dependency in a way that allows qmllint
to keep track of the accesses.

To help with allowing qmllint to have a defined shape to refer to, a new
custom QML type, `ListSelection` was added, whose only purpose is to
define a shape that contains the relevant property.

`State_Idle` was modified to work through an instance of that type,
providing a required property so that a reference to the instance can be
passed over and manipulated as required, with a shape that is known to
qmllint.

An instance of the new type was added as the root element of each list type and a
reference to the root element was passed over to the various `State_Idle`
instances that interact with it.

The inline definition of the various `list` property where removed from
the various instances of the relevant list types.

Pick-to: 6.10
Task-number: QTBUG-138170
Change-Id: I294a5c0f9f7175dd3706419d6fa02e7b3ed3cdb7
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>