FX&MaterialShowroom: Remove reference to parent in ui.qml file
ui.qml files do not support referencing the top-level `parent`, and
qmllint produces a warning when this is done.

Thus, remove the line that refers to `parent` from the ui.qml file.

In this particular case the only usage site of the type defined by the
relevant .ui.qml file sets the relevant property on the spot, such that
there should be no change in semantic with the removal.

Pick-to: 6.10
Task-number: QTBUG-138170
Change-Id: I68f9acc7c853afc1a62630a127bf1fa6eb891ec9
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>