FX&MaterialShowroom: Add a configuration file for qmlllint
The example is currently in a state where qmllint produces almost no
warnings, and should generally be kept in such a state.

Thus, add a configuration file that sets the warning limit to "0", such
that running the all_qmllint target for the example will fail if any
warning is produced, with the additional strictness hopefully helping
keeping the example almost warning-free.

Some warnings were disabled, instead of fixed, due to a compatibility
issue with Qt Design Studio, but should be fixed if that incompatibility
is resolved.

Fixes: QTBUG-138170
Change-Id: Ibbbf17892f6be1b7b549718df16036d6a57c3022
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>
(cherry picked from commit 6749ea650ee1ac94d12d14444b15fa5683ef29ec)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>