coffee example: Prepare to restore Qt 6.10 compatibility
We can't have cyclic dependencies between types in 6.10.

To reduce the cyclic dependency, we are extracting the coffee settings
into their own type.
The next commit will then (targeting only 6.10) replace ApplicationFlow
properties with var properies.
Reducing the coupling further by extracting the navigation functions
into yet another type is left as a future exercise.

Also, note the opportunity to fix the PropertyChanges in Progress.qml to
not reference a non-local id (which now wouldn't work anymore) – that
wasn't caught by qmllint because PropertyChanges is custom parsed, and
the only way to avoid this is to use the new syntax – but the warning
for that has been turned off due to QDS compatabililty.

Pick-to: dev
Task-number: QTBUG-138898
Change-Id: I6ae97a567350c98dcd8056ac8433858b0b8aec6f
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>